Update sponsorship email flags in blob storage, expirations
Blob storage keeps a [sponsorable]/[email] set of blobs that will be used to check for active sponsorships. Right now, no content is set for the blobs, just tags for easier searching afterwards.

Pending cancelation was replaced with the actual effective cancelled event, so we have to do less time-based checks, and also because undo of a pending cancellation doesn't result in another webhook call, so we miss that event and can improperly result in a cancellation.

The expirations are set in another table for one-time sponsorships, so the scheduler can check that. Also, we delete this expiration record when a one-time sponsorship becomes a montly one.