Improve diagnostics messages
The diagnostics reported via the source generator is rendered in VS with the wrong links and no description.

So we instead let an analyer pick up the messages from the generator (if any) and surface them as analyzer diagnostics which are far better supported in VS.

NOTE: if the analyzers are disabled, the source generator will report its own diagnostics, which will be lower "quality" but still preserve the right behavior.

Finally, the targets take care of detecting the full disabling via /nowarn or <NoWarn> of the generated diagnostics, and will render them via plan MSBuild, with no code so they cannot be disabled, with a text reminding the user that it's far better to contribute to the project instead :)