Introduce 12hrs CDN-based caching to improve performance
This should help in countries with slower backbone, and also potentially downstream ISP caching.

The CDN is configured to honor query strings in the caching, so we don't loose important telemetry via blob storage logs. We lose a bit of the real-time data from before, but this is a good compromise to get a better end-user experience.