Protect user personal data by hashing their email
Hash the email on the client before checking for app/sponsor status. This prevents this data from being PII from the get-go, and instead moves that knowledge only to *after* the explicit installation of the GH app which gives explicit consent to access the users's information about their email addresses.