Make SponsorLink check incremental
We want the pauses from this check to only happen incrementally, whenver the project or its dependencies change.

It turns out that Roslyn doesn't handle this for us as part of an incremental source generator, so we don't need to use that at all. So instead the check is now purely a diagnostic analyzer. We keep the previous API for compatibility only.