Improve support for packing targets with new simplified approach
The consuming package will now need to import the SL targets explicitly, as well as pack them, since there won't be a public dependency on it anymore in downstream package consumers.

So we move all relevant properties to a single targets file for simplicity, and the .props file now provides the `$(SponsorLinkTargets)` property to help packing. For now, we add NuGetizer support, but will need SDK Pack too.